관심사 분리하기
- Container에서 useEffect를 통해 상태체크를 하던 부분 제거
  (action을 통해 dispatch만 하도록 변경)
- api.js fetchRestaurants 예외처리 추가
- reducer에서 loadRestaurants 액션 및 api 호출 테스트하도록 추가
